Handle being called from a subdirectory
When called from within a subdirectory of the app (e.g. from within the
processing of buffalo test -m TestName) we need to strip an extra
directory off to be able to find the configs, etc.

Failing to do this results in errors like sqlite buildtags not being
applied correctly.

Fixes gobuffalo/buffalo#2284